% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasurementDiagnostics-package.R
\docType{package}
\name{MeasurementDiagnostics-package}
\alias{MeasurementDiagnostics}
\alias{MeasurementDiagnostics-package}
\title{MeasurementDiagnostics: Diagnostics for Lists of Codes Based on Measurements}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Diagnostics of list of codes based on concepts from the domains measurement and observation. This package works for data mapped to the Observational Medical Outcomes Partnership Common Data Model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/MeasurementDiagnostics/}
  \item Report bugs at \url{https://github.com/ohdsi/MeasurementDiagnostics/issues}
}

}
\author{
\strong{Maintainer}: Nuria Mercade-Besora \email{nuria.mercadebesora@ndorms.ox.ac.uk} (\href{https://orcid.org/0009-0006-7948-3747}{ORCID})

Authors:
\itemize{
  \item Edward Burn \email{edward.burn@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-9286-1128}{ORCID})
  \item Marta Alcalde-Herraiz \email{marta.alcaldeherraiz@ndorms.ox.ac.uk} (\href{https://orcid.org/0009-0002-4405-1814}{ORCID})
}

}
\keyword{internal}
