% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockMeasurementDiagnostics.R
\name{mockMeasurementDiagnostics}
\alias{mockMeasurementDiagnostics}
\title{Function to create a mock cdm reference.}
\usage{
mockMeasurementDiagnostics(
  nPerson = 100,
  con = DBI::dbConnect(duckdb::duckdb()),
  writeSchema = "main",
  seed = 111
)
}
\arguments{
\item{nPerson}{number of people in the cdm.}

\item{con}{A DBI connection to create the cdm mock object.}

\item{writeSchema}{Name of an schema on the same connection with writing
permissions.}

\item{seed}{seed to use when creating the mock data.}
}
\value{
cdm object
}
\description{
Creates an example dataset that can be used to show how
the package works
}
\examples{
\donttest{
library(MeasurementDiagnostics)

cdm <- mockMeasurementDiagnostics()

cdm
}
}
