% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{plotDoc}
\alias{plotDoc}
\title{Helper for consistent documentation of `plot`.}
\arguments{
\item{x}{Columns to use as horizontal axes. See options with
`visOmopResults::plotColumns(result)`.}

\item{y}{Columns to use as horizontal axes. See options with
`visOmopResults::plotColumns(result)`.}

\item{facet}{Columns to facet by. See options with
`visOmopResults::plotColumns(result)`. Formula input is also allowed to
specify rows and columns.}

\item{colour}{Columns to color by. See options with
`visOmopResults::plotColumns(result)`.}
}
\description{
Helper for consistent documentation of `plot`.
}
\keyword{internal}
