% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMeasurementTimings.R
\name{plotMeasurementTimings}
\alias{plotMeasurementTimings}
\title{Plot summariseMeasurementTiming results.}
\usage{
plotMeasurementTimings(
  result,
  x = "concept_id",
  plotType = "boxplot",
  timeScale = "days",
  facet = visOmopResults::strataColumns(result),
  colour = "cdm_name"
)
}
\arguments{
\item{x}{Columns to use as x axes.}

\item{plotType}{Type of desired formatted table, possibilities are "boxplot" and
"densityplot".}

\item{timeScale}{Time scale to show, it can be "days" or "years".}

\item{facet}{Columns to facet by. See options with
`availablePlotColumns(result)`. Formula is also allowed to specify rows and
columns.}

\item{colour}{Columns to color by. See options with
`availablePlotColumns(result)`.}
}
\value{
A ggplot.
}
\description{
Plot summariseMeasurementTiming results.
}
\examples{
\dontrun{
library(MeasurementDiagnostics)

cdm <- mockMeasurementDiagnostics()

result <- summariseMeasurementUse(
              cdm = cdm,
              codes = list("test_codelist" = c(3001467L, 45875977L)))
plotMeasurementTimings(result)

cdmDisconnect(cdm)
}

}
