% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseCohortMeasurementUse.R
\name{summariseCohortMeasurementUse}
\alias{summariseCohortMeasurementUse}
\title{Diagnostics of a codelist of measurement codes within a cohort}
\usage{
summariseCohortMeasurementUse(codes, cohort, timing = "during")
}
\arguments{
\item{codes}{A codelist of measurement codes for which to perform diagnostics.}

\item{cohort}{A cohort in which to perfom the diagnostics of the measurement
codes provided.}

\item{timing}{Three options: 1) "any" if the interest is on measurement
recorded any time, 2) "during", if interested in measurements while the
subject is in the cohort (or in observation if cohort = NULL), and 3)
"cohort_start_date" for measurements ocurring at cohort start date (or at
"observation_period_start_date if cohort = NULL).}
}
\value{
A summarised result
}
\description{
Diagnostics of a codelist of measurement codes within a cohort
}
\examples{
\donttest{
library(MeasurementDiagnostics)

cdm <- mockMeasurementDiagnostics()

result <- summariseCohortMeasurementUse(
  codes = list("test_codelist" = c(3001467L, 45875977L)),
  cohort = cdm$my_cohort, timing = "cohort_start_date"
)

CDMConnector::cdmDisconnect(cdm = cdm)
}

}
