% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseMeasurementUse.R
\name{summariseMeasurementUse}
\alias{summariseMeasurementUse}
\title{Diagnostics of a codelist of measurement codes in the database}
\usage{
summariseMeasurementUse(cdm, codes)
}
\arguments{
\item{cdm}{A reference to the cdm object.}

\item{codes}{A codelist of measurement codes for which to perform diagnostics.}
}
\value{
A summarised result
}
\description{
Diagnostics of a codelist of measurement codes in the database
}
\examples{
\donttest{
library(MeasurementDiagnostics)

cdm <- mockMeasurementDiagnostics()

result <- summariseMeasurementUse(
  cdm = cdm, codes = list("test_codelist" = c(3001467L, 45875977L))
)

CDMConnector::cdmDisconnect(cdm = cdm)
}

}
