% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableMeasurementTimings.R
\name{tableMeasurementTimings}
\alias{tableMeasurementTimings}
\title{Format a measurement_timings object into a visual table}
\usage{
tableMeasurementTimings(
  result,
  type = "gt",
  header = c(visOmopResults::strataColumns(result)),
  groupColumn = c("codelist_name"),
  hide = c("variable_name", "variable_level", "cohort_table"),
  .options = list()
)
}
\arguments{
\item{result}{A summarised_result object.}

\item{type}{Type of table. Check supported types with
`visOmopResults::tableType()`.}

\item{header}{Columns to use as header. See options with
`availableTableColumns(result)`.}

\item{groupColumn}{Columns to group by. See options with
`availableTableColumns(result)`.}

\item{hide}{Columns to hide from the visualisation. See options with
`availableTableColumns(result)`.}

\item{.options}{A named list with additional formatting options.
`visOmopResults::tableOptions()` shows allowed arguments and their default values.}
}
\value{
A formatted table
}
\description{
Format a measurement_timings object into a visual table
}
\examples{
\donttest{
library(MeasurementDiagnostics)

cdm <- mockMeasurementDiagnostics()
result <- summariseMeasurementUse(
              cdm = cdm,
              codes = list("test_codelist" = c(3001467L, 45875977L)))

tableMeasurementTimings(result)

CDMConnector::cdmDisconnect(cdm = cdm)
}

}
