% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{timeScaleDoc}
\alias{timeScaleDoc}
\title{Helper for consistent documentation of `timeScale`.}
\arguments{
\item{timeScale}{Time scale to show, it can be "days" or "years".}
}
\description{
Helper for consistent documentation of `timeScale`.
}
\keyword{internal}
