% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{uniqueCombinationsDoc}
\alias{uniqueCombinationsDoc}
\title{Helper for consistent documentation of `uniqueCombinations`.}
\arguments{
\item{uniqueCombinations}{Whether to restrict to unique reference and
comparator comparisons.}
}
\description{
Helper for consistent documentation of `uniqueCombinations`.
}
\keyword{internal}
