% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockMeasurementDiagnostics.R
\name{mockMeasurementDiagnostics}
\alias{mockMeasurementDiagnostics}
\title{Function to create a mock cdm reference.}
\usage{
mockMeasurementDiagnostics(
  nPerson = 100,
  source = "local",
  con = lifecycle::deprecated(),
  writeSchema = lifecycle::deprecated(),
  seed = lifecycle::deprecated()
)
}
\arguments{
\item{nPerson}{number of people in the cdm.}

\item{source}{The source where the cdm_reference object is inserted. Choice
between 'local' or 'duckdb'.}

\item{con}{deprecated.}

\item{writeSchema}{deprecated.}

\item{seed}{deprecated.}
}
\value{
cdm_reference object
}
\description{
Creates an example dataset that can be used to show how
the package works
}
\examples{
\donttest{
library(MeasurementDiagnostics)

cdm <- mockMeasurementDiagnostics(source = "duckdb")

cdm
}
}
