% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockMeasurementDiagnostics.R
\name{mockMeasurementDiagnostics}
\alias{mockMeasurementDiagnostics}
\title{Function to create a mock cdm reference.}
\usage{
mockMeasurementDiagnostics(nPerson = 100, recordPerson = 2, seed = 111)
}
\arguments{
\item{nPerson}{Number of people in the cdm.}

\item{recordPerson}{Average number of measurement records per person.}

\item{seed}{Seed to use when creating the mock data.}
}
\value{
A cdm object in a temporary duckdb database
}
\description{
Creates an example dataset that can be used to show how
the package works
}
\examples{
\donttest{
library(MeasurementDiagnostics)
cdm <- mockMeasurementDiagnostics()
cdm
}
}
