% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{plotDoc}
\alias{plotDoc}
\title{Helper for consistent documentation of `plot`.}
\arguments{
\item{facet}{Columns to facet by. See options with
`availablePlotColumns(result)`. Formula is also allowed to specify rows and
columns.}

\item{colour}{Columns to color by. See options with
`availablePlotColumns(result)`.}
}
\description{
Helper for consistent documentation of `plot`.
}
\keyword{internal}
