% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMeasurementValueAsConcept.R
\name{plotMeasurementValueAsConcept}
\alias{plotMeasurementValueAsConcept}
\title{Plot summariseMeasurementTiming results.}
\usage{
plotMeasurementValueAsConcept(
  result,
  x = "count",
  y = "codelist_name",
  facet = c("cdm_name"),
  colour = c("concept_name", "variable_level", visOmopResults::strataColumns(result)),
  style = NULL
)
}
\arguments{
\item{result}{A summarised_result object.}

\item{x}{Columns to use as horizontal axes. See options with
`visOmopResults::plotColumns(result)`.}

\item{y}{Columns to use as horizontal axes. See options with
`visOmopResults::plotColumns(result)`.}

\item{facet}{Columns to facet by. See options with
`visOmopResults::plotColumns(result)`. Formula input is also allowed to
specify rows and columns.}

\item{colour}{Columns to color by. See options with
`visOmopResults::plotColumns(result)`.}

\item{style}{Pre-defined style to apply: "default" or
"darwin" - the latter just for gt and flextable. If NULL the "default" style
is used.}
}
\value{
A ggplot.
}
\description{
Plot summariseMeasurementTiming results.
}
\examples{
\donttest{
library(MeasurementDiagnostics)
cdm <- mockMeasurementDiagnostics(source = "duckdb")
result <- summariseMeasurementUse(
              cdm = cdm,
              bySex = TRUE,
              codes = list("test_codelist" = c(3001467L, 45875977L)))
plotMeasurementValueAsConcept(result)
CDMConnector::cdmDisconnect(cdm)
}
}
