% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMeasurementValueAsConcept.R
\name{plotMeasurementValueAsConcept}
\alias{plotMeasurementValueAsConcept}
\title{Plot summariseMeasurementTiming results.}
\usage{
plotMeasurementValueAsConcept(
  result,
  x = "count",
  y = "variable_level",
  facet = c("cdm_name", "codelist_name", "concept_name"),
  colour = visOmopResults::strataColumns(result)
)
}
\arguments{
\item{result}{A summarised_result object.}

\item{x}{Columns to use as horizontal axes. See options with
`visOmopResults::plotColumns(result)`.}

\item{y}{Columns to use as horizontal axes. See options with
`visOmopResults::plotColumns(result)`.}

\item{facet}{Columns to facet by. See options with
`visOmopResults::plotColumns(result)`. Formula input is also allowed to
specify rows and columns.}

\item{colour}{Columns to color by. See options with
`visOmopResults::plotColumns(result)`.}
}
\value{
A ggplot.
}
\description{
Plot summariseMeasurementTiming results.
}
\examples{
\dontrun{
library(MeasurementDiagnostics)

cdm <- mockMeasurementDiagnostics()

result <- summariseMeasurementUse(
              cdm = cdm,
              bySex = TRUE,
              codes = list("test_codelist" = c(3001467L, 45875977L)))
plotMeasurementValueAsConcept(result)

cdmDisconnect(cdm)
}
}
