% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{tableDoc}
\alias{tableDoc}
\title{Helper for consistent documentation of `table`.}
\arguments{
\item{type}{Type of table. Check supported types with
`visOmopResults::tableType()`.}

\item{header}{Columns to use as header. See options with
`availableTableColumns(result)`.}

\item{groupColumn}{Columns to group by. See options with
`availableTableColumns(result)`.}

\item{hide}{Columns to hide from the visualisation. See options with
`availableTableColumns(result)`.}

\item{.options}{A named list with additional formatting options.
`visOmopResults::tableOptions()` shows allowed arguments and their default values.}
}
\description{
Helper for consistent documentation of `table`.
}
\keyword{internal}
